Fixed image update on repair, effect improvements
Building images were incorrectly not updating immediately when repaired - fixed with an update image bundle when a building is repaired. Additionally, the list of effects was identical in both configuration files, so it was condensed to a single effect list and 2 different active effect lists, allowing new effects to be added by only modifying 1 location.