Button constructor refactoring
As shown with the system used in actor constructors, it is easier to make configurable objects with optional parameters if constructors utilize an input dictionary rather than specifying all inputs as separate parameters - this commit partially implements the change with traditional buttons and label buttons, but does not yet change labels, label variants, or other button subclass families. This change is conducive to an interface collection system modeled off of the image bundle system.