Added armor, hats, minister portrait refactor
The minister portrait sections setup was moved to another function, as it was increasing in length and was unreliant on surrounding constructor code. Additionally, armor was added for aristocrat and royal hair ministers, always being worn by royal heirs and 50% of aristocrats. Hats were also added, with 2/3 of ministers having no hats - there are currently only armored hats